fix(lint): satisfy clippy on the pinned 1.96.0 toolchain
Pre-existing clippy debt that made `just clippy` red independent of any single
change (the gate stops at the first failing crate, so these surface one at a
time). None are behaviour changes:

- carrick-hal `vcpu_sched::global()`, carrick-vmm-hvf `save_guest_state`, and
  carrick-cli `build.rs`: intentional startup-invariant / fatal / build-time
  panics get an explicit `#[allow]` with rationale (the workspace denies
  expect/panic, so deliberate ones must be marked).
- carrick-thread `fork_quiesce`: a paragraph after a doc list read as a lazy
  continuation — separated with a blank doc line.
- carrick-runtime: drop three redundant `i32 -> i32` casts on M:N probe args
  and factor a 7-tuple epoll-interest type behind a `GfdInterest` alias.

`cargo clippy --workspace --all-targets --keep-going -- -D warnings` is clean.

+++ crates/carrick-runtime/src/vcpu_loop/threads.rs
@@ -91,7 +91,7 @@ where
                 }
                 let prev = old_slot.unwrap_or(new.slot);
                 crate::probes::mn_reclaim(
-                    self.this_tid as i32,
+                    self.this_tid,
                     prev,
                     new.slot,
                     if new.slot == prev { 1 } else { 2 },
@@ -99,7 +99,7 @@ where
             } else if engine.reclaims() {
                 // A reclaiming backend that PARKED (uncontended — kept its vCPU).
                 let slot = carrick_hal::vcpu_sched::current_slot().unwrap_or(0);
-                crate::probes::mn_reclaim(self.this_tid as i32, slot, slot, 0);
+                crate::probes::mn_reclaim(self.this_tid, slot, slot, 0);
             }
             let outcome = match raw {
                 FutexWaitOutcome::Woken => 0,
@@ -210,7 +210,7 @@ where
                 let lease = carrick_hal::vcpu_sched::global().acquire(tid as u64);
                 carrick_hal::vcpu_sched::set_current_lease(lease);
                 crate::probes::mn_admit(
-                    tid as i32,
+                    tid,
                     lease.slot,
                     carrick_hal::vcpu_sched::global()
                         .budget()
